Move NS state to separate submodule
